requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file dev.txt dev.in
#
asn1crypto==0.22.0        # via cryptography
attrs==17.2.0             # via automat, service-identity
automat==0.6.0            # via twisted
cffi==1.10.0              # via cryptography
constantly==15.1.0        # via twisted
coverage==4.4.1
cryptography==2.0.3       # via pyopenssl
cssselect==1.0.1          # via parsel, scrapy
decorator==4.1.1          # via ipython, traitlets
dj-database-url==0.4.2
django-getenv==1.3.2
django-money==0.11.4
django==1.11.3
factory-boy==2.9.2
faker==0.7.18             # via factory-boy
flake8==3.3.0
hyperlink==17.3.1         # via twisted
idna==2.6                 # via cryptography
incremental==17.5.0       # via twisted
ipdb==0.10.3
ipython-genutils==0.2.0   # via traitlets
ipython==5.4.1            # via ipdb
lxml==3.8.0               # via parsel, scrapy
mccabe==0.6.1             # via flake8
parsel==1.2.0             # via scrapy
pexpect==4.2.1            # via ipython
pickleshare==0.7.4        # via ipython
prompt-toolkit==1.0.14    # via ipython
psycopg2==2.7.1
ptyprocess==0.5.2         # via pexpect
py-moneyed==0.7.0         # via django-money
py==1.4.34                # via pytest
pyasn1-modules==0.0.11    # via service-identity
pyasn1==0.3.2             # via pyasn1-modules, service-identity
pycodestyle==2.3.1        # via flake8
pycparser==2.18           # via cffi
pydispatcher==2.0.5       # via scrapy
pyflakes==1.5.0           # via flake8
pygments==2.2.0           # via ipython
pyopenssl==17.2.0         # via scrapy, service-identity
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-mock==1.6.3
pytest==3.1.3
python-dateutil==2.6.1    # via faker
pytz==2017.2              # via django
queuelib==1.4.2           # via scrapy
scrapy==1.4.0
service-identity==17.0.0  # via scrapy
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via automat, cryptography, faker, parsel, prompt-toolkit, pyopenssl, python-dateutil, scrapy, traitlets, w3lib
traitlets==4.3.2          # via ipython
twisted==17.5.0           # via scrapy
uwsgi==2.0.15
w3lib==1.18.0             # via parsel, scrapy
wcwidth==0.1.7            # via prompt-toolkit
zope.interface==4.4.2     # via twisted
